Fix for effects filechooser location box text.
On first press of the effects configuration button the location box wasn't set so pressing 'Ok' would wipe the button's file setting. Presumably the problem is related to initialisation occurring outside of the gtk mainloop.

A neat side effect of the fix is that after issuing a 'Clear' a click on the config button results in the old file being preselected. This piece of data rightly does not survive application restart.